# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tauco Cap Badak API
  version: 1.1.0
  description: |
    Contract-first REST API for the Phase 1B public backend and the Phase 1C
    local shadow CMS. Administration operations are contractual only in C1 and
    are not registered by the runtime until their implementation gate.
  license:
    name: Proprietary
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
paths:
...
  /api/v1/admin/pages/{key}:
    get:
      tags: [Admin content]
      summary: Get Home or About with its revision pointers
      operationId: adminGetPage
      security:
        - AdminCookie: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/PageKeyPath'
      responses:
        '200':
          $ref: '#/components/responses/AdminPageOK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/pages/{key}/drafts:
    post:
      tags: [Admin content]
      summary: Save an immutable Home or About draft revision
      operationId: adminCreatePageDraft
      security:
        - AdminCookie: []
          AdminCSRF: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/PageKeyPath'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPageDraftRequest'
      responses:
        '201':
          $ref: '#/components/responses/AdminRevisionCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/pages/{key}/revisions/{revisionId}:
    get:
      tags: [Admin content]
      summary: Get one immutable page revision
      operationId: adminGetPageRevision
      security:
        - AdminCookie: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/PageKeyPath'
        - $ref: '#/components/parameters/RevisionID'
      responses:
        '200':
          $ref: '#/components/responses/AdminRevisionOK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/pages/{key}/revisions/{revisionId}/publish:
    post:
      tags: [Admin content]
      summary: Publish a new immutable snapshot from a page revision
      operationId: adminPublishPageRevision
      security:
        - AdminCookie: []
          AdminCSRF: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/PageKeyPath'
        - $ref: '#/components/parameters/RevisionID'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          $ref: '#/components/responses/AdminRevisionOK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/pages/{key}/unpublish:
    post:
      tags: [Admin content]
      summary: Clear a page public pointer without deleting revisions
      operationId: adminUnpublishPage
      security:
        - AdminCookie: []
          AdminCSRF: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/PageKeyPath'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          $ref: '#/components/responses/AdminNoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    InternalMetricsBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: A dedicated rotatable operations token; never an admin JWT.
    AdminCookie:
      type: apiKey
      in: cookie
      name: tauco_admin_access
      description: Short-lived HttpOnly, Secure, SameSite=Strict admin access cookie.
    AdminCSRF:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-submit token required with AdminCookie on unsafe methods.
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      description: |
        Optional caller-provided trace identifier. Invalid values are replaced
        by the server.
      schema:
        $ref: '#/components/schemas/RequestID'
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag for optimistic concurrency control.
      schema:
        type: string
        minLength: 3
        maxLength: 160
      example: '"revision-7"'
    PageKeyPath:
      name: key
      in: path
      required: true
      schema:
        type: string
        enum: [home, about]
...
    RevisionID:
      name: revisionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    XRequestID:
      description: Request trace identifier.
      required: true
      schema:
        $ref: '#/components/schemas/RequestID'
    ETag:
      description: Strong validator derived from the immutable revision checksum.
      required: true
      schema:
        type: string
      example: '"sha256-8e7d6f5c4b3a"'
...
    NoStore:
      description: Prevents storage of private or operational responses.
      required: true
      schema:
        type: string
        enum:
          - no-store
...
  responses:
...
    BadRequest:
      description: Malformed JSON, header, cursor, or query
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            malformedRequest:
              value:
                type: urn:tauco-cap-badak:problem:bad-request
                title: Permintaan tidak valid
                status: 400
                detail: Format permintaan tidak dapat diproses.
                instance: /api/v1/contact-messages
                code: BAD_REQUEST
                requestId: 98a70ee8a919423b883c23ea275d468a
...
    ValidationFailed:
      description: One or more semantic field validations failed
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            invalidEmail:
              value:
                type: urn:tauco-cap-badak:problem:validation
                title: Permintaan tidak valid
                status: 422
                detail: Periksa field yang ditandai.
                instance: /api/v1/contact-messages
                code: CONTACT_VALIDATION_FAILED
                requestId: 98a70ee8a919423b883c23ea275d468a
                errors:
                  - field: email
                    code: INVALID_EMAIL
                    message: Masukkan alamat email yang valid.
...
    AdminPageOK:
      description: Editable page and revision pointers
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        ETag: { $ref: '#/components/headers/ETag' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminPageResponse' }
    AdminRevisionOK:
      description: Immutable content revision
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        ETag: { $ref: '#/components/headers/ETag' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminRevisionResponse' }
    AdminRevisionCreated:
      description: Immutable draft revision created
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        ETag: { $ref: '#/components/headers/ETag' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminRevisionResponse' }
...
    AdminNoContent:
      description: Operation completed with no response body
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
...
    NotFound:
      description: Requested administration resource does not exist
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Current resource state conflicts with the requested operation
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionFailed:
      description: If-Match no longer identifies the current base revision
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid operations token
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema:
            type: string
          example: Bearer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            unauthorized:
              value:
                type: urn:tauco-cap-badak:problem:unauthorized
                title: Autentikasi diperlukan
                status: 401
                detail: Token operasional tidak tersedia atau tidak valid.
                instance: /internal/metrics
                code: UNAUTHORIZED
                requestId: 98a70ee8a919423b883c23ea275d468a
    Forbidden:
      description: Authenticated principal is not allowed to read metrics
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            forbidden:
              value:
                type: urn:tauco-cap-badak:problem:forbidden
                title: Akses ditolak
                status: 403
                detail: Kredensial tidak memiliki akses ke resource ini.
                instance: /internal/metrics
                code: FORBIDDEN
                requestId: 98a70ee8a919423b883c23ea275d468a
    InternalServerError:
      description: Unexpected server error with no internal detail exposed
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            internal:
              value:
                type: urn:tauco-cap-badak:problem:internal
                title: Terjadi kesalahan internal
                status: 500
                detail: Permintaan tidak dapat diproses saat ini.
                instance: /api/v1/home
                code: INTERNAL_SERVER_ERROR
                requestId: 98a70ee8a919423b883c23ea275d468a
...
  schemas:
    RequestID:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9._:-]+$'
      example: 98a70ee8a919423b883c23ea275d468a
...
    SeoMetadata:
      type: object
      additionalProperties: false
      required:
        - title
        - description
        - canonicalPath
        - openGraphImage
      properties:
        title:
          type: string
          minLength: 8
          maxLength: 80
          pattern: '^\S(?:[\s\S]*\S)?$'
        description:
          type: string
          minLength: 50
          maxLength: 180
          pattern: '^\S(?:[\s\S]*\S)?$'
        canonicalPath:
          $ref: '#/components/schemas/InternalPath'
        openGraphImage:
          $ref: '#/components/schemas/InformativeImageAsset'
    TextSection:
      type: object
      additionalProperties: false
      required:
        - id
        - heading
        - paragraphs
      properties:
        id:
          $ref: '#/components/schemas/Slug'
        heading:
          type: string
          minLength: 2
          maxLength: 120
          pattern: '^\S(?:[\s\S]*\S)?$'
        paragraphs:
          type: array
          minItems: 1
          maxItems: 6
          items:
            type: string
            minLength: 20
            maxLength: 600
            pattern: '^\S(?:[\s\S]*\S)?$'
...
    HomeHero:
      type: object
      additionalProperties: false
      required:
        - eyebrow
        - title
        - description
        - image
        - actions
      properties:
        eyebrow:
          type: string
          minLength: 2
          maxLength: 120
          pattern: '^\S(?:[\s\S]*\S)?$'
        title:
          type: string
          minLength: 4
          maxLength: 100
          pattern: '^\S(?:[\s\S]*\S)?$'
        description:
          type: string
          minLength: 40
          maxLength: 240
          pattern: '^\S(?:[\s\S]*\S)?$'
        image:
          $ref: '#/components/schemas/ImageAsset'
        actions:
          type: array
          minItems: 1
          maxItems: 2
          items:
            $ref: '#/components/schemas/InternalLink'
...
    HomeContent:
      type: object
      additionalProperties: false
      required:
        - metadata
        - hero
        - introduction
        - featuredProductSlugs
        - guidePreview
        - aboutPreview
      properties:
        metadata:
          $ref: '#/components/schemas/SeoMetadata'
        hero:
          $ref: '#/components/schemas/HomeHero'
        introduction:
          $ref: '#/components/schemas/TextSection'
        featuredProductSlugs:
          type: array
          maxItems: 6
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Slug'
        guidePreview:
          $ref: '#/components/schemas/ContentPreview'
        aboutPreview:
          $ref: '#/components/schemas/ContentPreview'
    AboutContent:
      type: object
      additionalProperties: false
      required:
        - metadata
        - hero
        - sections
        - relatedLinks
        - sources
      properties:
        metadata:
          $ref: '#/components/schemas/SeoMetadata'
        hero:
          $ref: '#/components/schemas/PageHero'
        sections:
          type: array
          minItems: 2
          maxItems: 8
          items:
            $ref: '#/components/schemas/TextSection'
        relatedLinks:
          type: array
          minItems: 1
          maxItems: 4
          items:
            $ref: '#/components/schemas/InternalLink'
        sources:
          type: array
          minItems: 1
          maxItems: 8
          items:
            $ref: '#/components/schemas/SourceReference'
...
    ResponseMeta:
      type: object
      additionalProperties: false
      required:
        - requestId
        - apiVersion
      properties:
        requestId:
          $ref: '#/components/schemas/RequestID'
        apiVersion:
          type: string
          enum:
            - v1
...
    ValidationError:
      type: object
      additionalProperties: false
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 120
          pattern: '^\S(?:[\s\S]*\S)?$'
        code:
          type: string
          minLength: 3
          maxLength: 80
          pattern: '^[A-Z][A-Z0-9]*(?:_[A-Z0-9]+)*$'
        message:
          type: string
          minLength: 3
          maxLength: 240
          pattern: '^\S(?:[\s\S]*\S)?$'
    Problem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
        - instance
        - code
        - requestId
      properties:
        type:
          type: string
          format: uri
          maxLength: 160
          pattern: '^urn:tauco-cap-badak:problem:[a-z0-9]+(?:-[a-z0-9]+)*$'
        title:
          type: string
          minLength: 3
          maxLength: 120
          pattern: '^\S(?:[\s\S]*\S)?$'
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 3
          maxLength: 500
          pattern: '^\S(?:[\s\S]*\S)?$'
        instance:
          $ref: '#/components/schemas/ProblemInstance'
        code:
          type: string
          minLength: 3
          maxLength: 80
          pattern: '^[A-Z][A-Z0-9]*(?:_[A-Z0-9]+)*$'
        requestId:
          $ref: '#/components/schemas/RequestID'
        errors:
          type: array
          minItems: 1
          maxItems: 32
          items:
            $ref: '#/components/schemas/ValidationError'
    ProblemInstance:
      type: string
      format: uri-reference
      minLength: 1
      maxLength: 2048
      pattern: "^/(?:$|(?:[A-Za-z0-9._~!$&'()*+,;=:@-]|%[0-9A-Fa-f]{2})(?:[A-Za-z0-9._~!$&'()*+,;=:@/-]|%[0-9A-Fa-f]{2})*)$"
      description: |
        Absolute, path-only URI-reference built from the escaped request path.
        Uppercase segments, a trailing slash, and percent-encoded octets are
        valid. Empty, non-absolute, network-path, or oversized values fall back
        to `/` before a problem response is emitted.
      x-source: escaped-request-path
...
    AdminPageContent:
      oneOf:
        - $ref: '#/components/schemas/HomeContent'
        - $ref: '#/components/schemas/AboutContent'
...
    AdminPageDraftRequest:
      type: object
      additionalProperties: false
      required: [baseRevisionId, content]
      properties:
        baseRevisionId:
          type: string
          format: uuid
        content:
          $ref: '#/components/schemas/AdminPageContent'
...
    AdminRevision:
      type: object
      additionalProperties: false
      required: [id, ownerId, revisionNumber, status, schemaVersion, content, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        revisionNumber:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [draft, published, archived]
        schemaVersion:
          type: integer
          minimum: 1
        content:
          $ref: '#/components/schemas/AdminRevisionContent'
        createdBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
          nullable: true
    AdminRevisionResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/AdminRevision' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    AdminRevisionSummary:
      type: object
      additionalProperties: false
      required: [id, revisionNumber, status, createdBy, createdAt, publishedAt]
      properties:
        id:
          type: string
          format: uuid
        revisionNumber:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [draft, published, archived]
        createdBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
          nullable: true
    AdminPage:
      type: object
      additionalProperties: false
      required: [id, key, latestRevision, revisions, publishedRevisionId, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
          enum: [home, about]
        latestRevision:
          $ref: '#/components/schemas/AdminRevision'
        revisions:
          type: array
          description: Latest 100 immutable revisions ordered newest first.
          maxItems: 100
          items: { $ref: '#/components/schemas/AdminRevisionSummary' }
        publishedRevisionId:
          type: string
          format: uuid
          nullable: true
        updatedAt:
          type: string
          format: date-time
    AdminPageResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/AdminPage' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
...